+++ Tests/MapboxNavigationTests/StatusViewTests.swift
@@ -11,14 +11,14 @@ class StatusViewTests: XCTestCase {
     }()
 
     func testWithDelayShorterThanDuration() {
-        addNewStatus(status: firstStatus())
+        show(firstStatus())
         XCTAssertEqual(self.statusView.statuses.count, 1)
     }
 
     func testWithDelayLongerThanDuration() {
         let seconds = 5.0
         XCTAssertTrue(statusView.isHidden)
-        addNewStatus(status: firstStatus())
+        show(firstStatus())
         XCTAssertFalse(statusView.isHidden)
         let path = #keyPath(UIView.isHidden)
         let expectation = XCTKVOExpectation(keyPath: path, object: statusView, expectedValue: true)
@@ -28,9 +28,9 @@ class StatusViewTests: XCTestCase {
     }
 
     func testFirstAndSecond() {
-        addNewStatus(status: firstStatus())
+        show(firstStatus())
         XCTAssertFalse(statusView.isHidden)
-        addNewStatus(status: secondStatus())
+        show(secondStatus())
         XCTAssertEqual(statusView.statuses.count, 2)
         let path = #keyPath(UIView.isHidden)
         let expectation = XCTKVOExpectation(keyPath: path, object: statusView, expectedValue: true)
@@ -40,8 +40,8 @@ class StatusViewTests: XCTestCase {
     }
     
     func testWithInfinite() {
-        addNewStatus(status: firstStatus())
-        addNewStatus(status: thirdStatus())
+        show(firstStatus())
+        show(thirdStatus())
         XCTAssertEqual(self.statusView.statuses.count, 2)
         let path = #keyPath(UIView.isHidden)
         let expectation = XCTKVOExpectation(keyPath: path, object: statusView, expectedValue: true)
@@ -55,26 +55,23 @@ extension StatusViewTests {
     
     // define statuses
     func firstStatus() -> StatusView.Status {
-        let title1 = NSLocalizedString("FIRST_TEST_STATUS", bundle: .mapboxNavigation, value: "first test status", comment: "the first status banner used for testing")
-        return StatusView.Status(id: title1, duration: 1, priority: StatusView.Priority(rawValue: 0))
+        return StatusView.Status(id: "first test status", duration: 1, priority: StatusView.Priority(rawValue: 0))
     }
     
     func secondStatus() -> StatusView.Status {
-        let title2 = NSLocalizedString("SECOND_TEST_STATUS", bundle: .mapboxNavigation, value: "second test status", comment: "the second status banner used for testing")
-        return StatusView.Status(id: title2, duration: 5, priority: StatusView.Priority(rawValue: 1))
+        return StatusView.Status(id: "second test status", duration: 5, priority: StatusView.Priority(rawValue: 1))
     }
     
     func thirdStatus() -> StatusView.Status {
-        let title3 = NSLocalizedString("THIRD_TEST_STATUS", bundle: .mapboxNavigation, value: "third test status", comment: "the third status banner used for testing")
-        return StatusView.Status(id: title3, duration: .infinity, priority: StatusView.Priority(rawValue: 2))
+        return StatusView.Status(id: "third test status", duration: .infinity, priority: StatusView.Priority(rawValue: 2))
     }
 
-    func addNewStatus(status: StatusView.Status) {
-        statusView.addNewStatus(status: status)
+    func show(_ status: StatusView.Status) {
+        statusView.show(status)
     }
     
-    func hideStatus(with status: StatusView.Status) {
-        statusView.hideStatus(using: status)
+    func hide(_ status: StatusView.Status) {
+        statusView.hide(status)
     }
     
     func clearStatuses() {